requirement.txt
Django>=4.2
djangorestframework>=3.13
requests>=2.28
python-decouple>=3.8
